charmap fixed (for real this time)
When I originally added the missing names to charmap.txt, I hadn't fetched the most recent version of the branch. Once I did that, I messed up the merge and left out NAME_MEDIUM, which was on the remote but not my local branch. I added that back in and now the rom builds again.