Player transform into Pokémon system (#260)
* chore(flags): add flag to toggle followers

* feat(overworld): implement player transform

Allows the player to transform into a pokemon in their party.
This will replace the player gfx with the follower mon gfx

* fix(flash): make flash circle center on transformed player

Flash radius will be determined by the pokemon the player is
currently transformed into.

* feat(overworld): persist transform data in saveblock

Adds 3 bits to SaveBlock1. Does not break save compatibility.

* fix(overworld): fix player gender default to male on transform back

Transforming back to human would always set the player avatar to the
male one

* feat(menu): hide some start menu entries when player is pokemon

* fix(overworld): fix follower glitch when transforming back

The follower mon sprite would glitch for a few frames when the player
transformed back to human. This only happened if the player was
previously transformed into the current follower mon. We circumvent this
by resetting the player avatar instead of changing the graphics id.

* feat(overworld): implement player avatar bobbing when transformed

* chore(transform): remove unused task function definition

* chore(transform): mark palette update func as unused

* feat(script): add transformplayer macro

The macro `transformplayer` will open the party menu if used without an
argument. Use `transformplayer human` or `transformplayer 6` to
transform the player back to human. Provide a party index as the
argument to transform to that pokemon.

* feat(script): add more functionality to transform macro

Adds option to defer the transformation until the player object is
reloaded. Also adds an option to read from a special var.

* fix(transform): fix minor typo in function name

* refactor(transform): add multiple macros for transform

Introduces the following macros:
- transformplayer (reads from VAR_0x8004)
- autotransformplayer (opens party menu)
- detransformplayer (turn back to human)
- transformplayerto (takes 1-indexed party slot as arg)

All except autotransformplayer can take defer as an arg. If set the
player avatar will not be reloaded immediately.

* fix(overworld): fix flash mon level if transformed

Uses the currently transformed pokemon's level instead of the follower
level if the player is transformed.

* fix(overworld): fix bobbing and movement during deferred transform